Fix and clean up
Former-commit-id: 8e656a8932ea6b9dea33bf6d464cc41420ee9961 [formerly b55401176dd70f7a7721ce94160b78767b7c8d60]
Former-commit-id: c7e426f5b61bae5ecab4e51900ff27fc2beca540 [formerly 92ba4d6e5116d5b29aeb53c73eca25e9c7ca18e0]
Former-commit-id: 8ed4cbd9e3c05f6d7ccd8e03e323823dd030cecb
Former-commit-id: 22ff9ddcf1fe3a78278833897480db2826b249bd

+++ Source/RawParser/Model/Parser/Decoder/Decompressor/NikonDecompressor.cs
@@ -111,7 +111,7 @@ namespace RawNet.Decoder.Decompressor
                 pUp2[y & 1] += HuffDecodeNikon(bits);
                 pLeft1 = pUp1[y & 1];
                 pLeft2 = pUp2[y & 1];
-                int dest = y * raw.raw.dim.width;
+                long dest = y * raw.raw.dim.width;
                 raw.SetWithLookUp((ushort)Common.Clampbits(pLeft1, 15), raw.raw.data, dest++, ref random);
                 raw.SetWithLookUp((ushort)Common.Clampbits(pLeft2, 15), raw.raw.data, dest++, ref random);
                 for (int x = 1; x < raw.raw.dim.width / 2; x++)
